Refactor cloud discovery into discovery package
