+++ pkg/cloudprovider/discovery/discovery.go
@@ -1,0 +1,74 @@
+/*
+Copyright The Kubernetes Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package discovery
+
+import (
+	"context"
+	"fmt"
+
+	"cloud.google.com/go/compute/metadata"
+	"sigs.k8s.io/dranet/pkg/cloudprovider"
+	"sigs.k8s.io/dranet/pkg/cloudprovider/aws"
+	"sigs.k8s.io/dranet/pkg/cloudprovider/azure"
+	"sigs.k8s.io/dranet/pkg/cloudprovider/gce"
+	"sigs.k8s.io/dranet/pkg/cloudprovider/oke"
+)
+
+type CloudProviderHint string
+
+const (
+	CloudProviderHintGCE     CloudProviderHint = "GCE"
+	CloudProviderHintAWS     CloudProviderHint = "AWS"
+	CloudProviderHintAzure   CloudProviderHint = "AZURE"
+	CloudProviderHintOKE     CloudProviderHint = "OKE"
+	CloudProviderHintNone    CloudProviderHint = "NONE"
+)
+
+// DiscoverCloudProvider probes the environment to detect which cloud provider DRANET is running on.
+func DiscoverCloudProvider(ctx context.Context) CloudProviderHint {
+	if metadata.OnGCE() {
+		return CloudProviderHintGCE
+	}
+	if aws.OnAWS(ctx) {
+		return CloudProviderHintAWS
+	}
+	if azure.OnAzure(ctx) {
+		return CloudProviderHintAzure
+	}
+	if oke.OnOKE(ctx) {
+		return CloudProviderHintOKE
+	}
+	return CloudProviderHintNone
+}
+
+// GetInstanceProperties initializes and returns the specified cloud provider instance.
+func GetInstanceProperties(ctx context.Context, hint CloudProviderHint) (cloudprovider.CloudInstance, error) {
+	switch hint {
+	case CloudProviderHintGCE:
+		return gce.GetInstance(ctx)
+	case CloudProviderHintAWS:
+		return aws.GetInstance(ctx)
+	case CloudProviderHintAzure:
+		return azure.GetInstance(ctx)
+	case CloudProviderHintOKE:
+		return oke.GetInstance(ctx)
+	case CloudProviderHintNone, "none", "":
+		return nil, nil
+	default:
+		return nil, fmt.Errorf("unknown cloud provider hint: %s", hint)
+	}
+}
