+++ pkg/inventory/db.go
@@ -95,11 +95,6 @@ type DB struct {
 	// When false, IPoIB interfaces are skipped and the underlying device is
 	// exposed as an IB-only RDMA device.
 	moveIBInterfaces bool
-
-	// cloudProviderHint is a hint for the cloud provider that will be used to
-	// select the appropriate provider plugin. Supported values: (GCE, AZURE, OKE, NONE).
-	// If not set, the cloud provider will be auto-detected, but may take longer to start.
-	cloudProviderHint CloudProviderHint
 }
 
 type Option func(*DB)
@@ -122,16 +117,9 @@ func WithMoveIBInterfaces(move bool) Option {
 	}
 }
 
-func WithCloudProviderHint(hint string) Option {
+func WithCloudInstance(instance cloudprovider.CloudInstance) Option {
 	return func(db *DB) {
-		if hint != "" {
-			// validate hint
-			h := CloudProviderHint(hint)
-			if h != CloudProviderHintGCE && h != CloudProviderHintAWS && h != CloudProviderHintAzure && h != CloudProviderHintOKE && h != CloudProviderHintNone {
-				klog.Fatalf("unknown cloud provider hint %q", hint)
-			}
-			db.cloudProviderHint = h
-		}
+		db.instance = instance
 	}
 }
 
@@ -165,10 +153,6 @@ func (db *DB) Run(ctx context.Context) error {
 	}
 
 	// Obtain data that will not change after the startup
-	if db.cloudProviderHint == "" {
-		db.cloudProviderHint = discoverCloudProvider(ctx)
-	}
-	db.instance = getInstanceProperties(ctx, db.cloudProviderHint)
 	db.gwInterfaces = getExcludedUplinkInterfaces()
 	klog.V(2).Infof("Excluded uplink interfaces and children: %v", db.gwInterfaces.UnsortedList())
 
@@ -529,7 +513,7 @@ func (db *DB) discoverRDMADevices(devices []resourceapi.Device) []resourceapi.De
 func (db *DB) addCloudAttributes(devices []resourceapi.Device) []resourceapi.Device {
 	for i := range devices {
 		device := &devices[i]
-		maps.Copy(device.Attributes, getProviderAttributes(device, db.instance))
+		maps.Copy(device.Attributes, db.getProviderAttributes(device, db.instance))
 	}
 	return devices
 }
@@ -671,4 +655,29 @@ func isAllocatableNetworkDevice(dev *ghw.PCIDevice) bool {
 		return false
 	}
 	return !nonNetdevDrivers.Has(dev.Driver)
+}
+
+// getProviderAttributes retrieves cloud provider-specific attributes for a network interface
+func (db *DB) getProviderAttributes(device *resourceapi.Device, instance cloudprovider.CloudInstance) map[resourceapi.QualifiedName]resourceapi.DeviceAttribute {
+	if instance == nil {
+		klog.Warningf("instance metadata is nil, cannot get provider attributes.")
+		return nil
+	}
+	if device == nil {
+		klog.Warningf("device is nil, cannot get provider attributes.")
+		return nil
+	}
+
+	id := cloudprovider.DeviceIdentifiers{
+		Name: device.Name,
+	}
+	// get the device identifiers from the device attributes
+	if macAttr, ok := device.Attributes[apis.AttrMac]; ok && macAttr.StringValue != nil {
+		id.MAC = *macAttr.StringValue
+	}
+	if pciAttr, ok := device.Attributes[apis.AttrPCIAddress]; ok && pciAttr.StringValue != nil {
+		id.PCIAddress = *pciAttr.StringValue
+	}
+
+	return instance.GetDeviceAttributes(id)
 }
